moved errordata out from contexts to model
