+++ frontend/src/model/ErrorData.ts
@@ -1,7 +1,3 @@
 export interface ErrorData {
   error: string;
 }
-
-// block
-// unblock
-// unfriend
